MRSHub Projects Page
Built a community Projects Page and included the Synthetic MRS Working Group Project.

The Community MRS projects are ongoing projects that involve multiple members of the MRS community and are open to collaborate with other spectroscopists.

The Projects page has been added to the navigation-bar. The redundant Home tab has been removed.
- Aaron